LayoutTests/fast/canvas/canvas-fill-rule-expected.txt

Series of tests to ensure correct results of the winding rule.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Testing default fill
PASS pixelDataAtPoint()[0] is within 5 of 0
PASS pixelDataAtPoint()[1] is within 5 of 255
PASS pixelDataAtPoint()[2] is within 5 of 0
PASS pixelDataAtPoint()[3] is within 5 of 255

Testing nonzero fill
PASS pixelDataAtPoint()[0] is within 5 of 0
PASS pixelDataAtPoint()[1] is within 5 of 255
PASS pixelDataAtPoint()[2] is within 5 of 0
PASS pixelDataAtPoint()[3] is within 5 of 255

Testing evenodd fill
PASS pixelDataAtPoint()[0] is within 5 of 255
PASS pixelDataAtPoint()[1] is within 5 of 0
PASS pixelDataAtPoint()[2] is within 5 of 0
PASS pixelDataAtPoint()[3] is within 5 of 255

PASS successfullyParsed is true

TEST COMPLETE
